refactor: Move exporter-only types to src/exporter/types.ts
